requirements.txt
aiofiles==23.1.0
aiogram==3.0.0b7
aiohttp==3.8.5
aiosignal==1.3.1
async-timeout==4.0.2
asyncio==3.4.3
attrs==23.1.0
Babel==2.9.1
certifi==2023.7.22
charset-normalizer==3.2.0
environs==9.5.0
frozenlist==1.4.0
idna==3.4
magic-filter==1.0.11
marshmallow==3.20.1
multidict==6.0.4
packaging==23.1
pydantic==1.10.12
python-dotenv==1.0.0
pytz==2023.3
typing_extensions==4.7.1
urllib3==2.0.4
yarl==1.9.2
